feat(TaskRecyclerAdapter): implement a RecyclerView Adapter class for the task list RecyclerView.
This class overrides the methods `onCreateViewHolder`, `onBindViewHolder` and `getItemCount` of the superclass. The class is used to inflate the task list view and bind the task data to the view items.

Additionally, an inner class `ViewHolder`, which extends RecyclerView.ViewHolder is also implemented here. The ViewHolder class is used to hold and handle the individual views within the RecyclerView. It can also be used to register event listeners for the view, or individual items within the view.