refactor(MainActivity): make new tasks appear at the top of the list, rather than at the bottom after being created.
This was done by adding new items in the ArrayList to the beginning of the list (by passing 0 as the index argument to `add()`). In the future, we could add sorting options to the settings, to sort the list by creation date, etc. But that would require having an actual `Task` class and object to store things like date created/updated, is task complete, etc.